Remove templace cache
This dramatically slows down the backend's performance.
Because images are stored in disk, and which can be
of very large size (e.g., 8mb), the speed for fetching
each image from the disk can be quite slow. Unlike streaming
videos, such latency is very noticable for images. The
template cache wastes backend's performance by requesting
images that are not displayed at all, turning it off kinda
solves this problem.